+++ examples/conditional-request-handling/JsonRequestHandler.cpp
@@ -3,7 +3,6 @@
 #include "JsonRequestHandler.h"
 #include <string>
 #include "yarpl/Flowable.h"
-#include "yarpl/Flowables.h"
 
 using namespace reactivesocket;
 using namespace rsocket;
