+++ examples/conditional-request-handling/TextRequestHandler.cpp
@@ -3,7 +3,6 @@
 #include "TextRequestHandler.h"
 #include <string>
 #include "yarpl/Flowable.h"
-#include "yarpl/Flowables.h"
 
 using namespace reactivesocket;
 using namespace rsocket;
