+++ examples/stream-hello-world/StreamHelloWorld_Client.cpp
@@ -11,7 +11,6 @@
 #include "rsocket/transports/TcpConnectionFactory.h"
 
 #include "yarpl/Flowable.h"
-#include "yarpl/flowable/Subscriber.h"
 
 using namespace reactivesocket;
 using namespace rsocket_example;
