+++ examples/stream-hello-world/StreamHelloWorld_Server.cpp
@@ -9,7 +9,6 @@
 #include "rsocket/RSocket.h"
 #include "rsocket/transports/TcpConnectionAcceptor.h"
 #include "yarpl/Flowable.h"
-#include "yarpl/Flowables.h"
 
 using namespace reactivesocket;
 using namespace rsocket;
