+++ examples/stream-observable-to-flowable/StreamObservableToFlowable_Client.cpp
@@ -11,7 +11,6 @@
 #include "rsocket/transports/TcpConnectionFactory.h"
 
 #include "yarpl/Flowable.h"
-#include "yarpl/flowable/Subscribers.h"
 
 using namespace reactivesocket;
 using namespace rsocket_example;
