+++ experimental/rsocket-src/RSocketRequester.cpp
@@ -4,7 +4,6 @@
 
 #include "rsocket/OldNewBridge.h"
 #include "yarpl/Flowable.h"
-#include "yarpl/Flowables.h"
 
 #include <folly/ExceptionWrapper.h>
