+++ experimental/yarpl/test/Observable_test.cpp
@@ -5,12 +5,9 @@
 #include <atomic>
 
 #include "yarpl/Observable.h"
-#include "yarpl/Observables.h"
-#include "yarpl/ThreadScheduler.h"
+#include "yarpl/schedulers/ThreadScheduler.h"
 #include "yarpl/flowable/Subscriber.h"
 #include "yarpl/flowable/Subscribers.h"
-#include "yarpl/observable/Observers.h"
-#include "yarpl/observable/Subscriptions.h"
 
 #include "Tuple.h"
